Moving some code around.
